split cache.rs into two files
